Remove users.files in favour of the new write_files[*].defer field
This change set implements the feedback from recent discussions [1] on how to best
support writing files at some later point during execution, so that, for example, a
user who is configured as file owner would exist when the file is being created.

THe general idea here is to keep the structure of the handlers of both modules
(write_files[_deferred]) as similar as possible, since they do pretty much the
same. The main difference is the lambda function used to filter deferred or not
deferred files.

For now, there is just some unit testing for the module handler. Integration tests
should be added after the implementation has been reviewed.

Specific details worth mentioning:
* 'validate_cloudconfig_schema' has been moved to the top of the handler functions,
  because
    a) even if the 'write_files' top-level directive is configured and not empty, it
       might still be empty for one of the two write-file modules
    b) the schema validation is a formal check and has no awareness of any logical
       semantics, so it makes sense to first verify the former before moving forward
* the new module handling deferred files has been renamed
* write_files.handle has been slightly restructured to adhere to the changes in
  write_files_deferred.handle
* revert all changes made to cc_users_groups.py in this branche while working on
  the deferred files feature

[1] https://github.com/canonical/cloud-init/pull/916